+++ frontend/src/components/CategoryOverview.svelte
@@ -5,6 +5,9 @@
 	let categoryMarkedForDeletionNameTypedByUser;
 </script>
 
+
+
+
 {#each $categoryStore as category}
 	<div>
 		{category.name}
